Read kernel into allocated pages (Bug fix)
Fixes bug where we were loading the kernel into the virtual address
specified by the ELF file instead of acutal allocated memory. The kernel
only accidentally worked before this point.

We make one big call to AllocatePages and manually
load the kernel segments at the correct addresses.

Changes had to be made to gdt.c and idt.c because the compiler was
optimizing in the incorrect virtuall address in the ELF file instead of
taking the acutal runtime address.