Add frame allocator
The frame allocator allocates blocks of physical memory to be used by
allocators which then maps this memory into linear memory.

The allocator uses a buddy allocation system with the min size block
being equal to the PAGE_SIZE and the max order being 7. Blocks are
currently merged when freeing.

There are two flavours of allocations functions, one for normal
allocation and one for contiguos physical memory which is required by
DMA.

The contiguos allocation function is kind of temporary and can currently
only allocate blocks of exact order sizes.