Move entry pool into seperate file
It's move so that the pool can be used by additional memory systems like
the paging system.

This commit also changes the behavior of the entry pool,
it now allocates more entries when a threshold is reached.
We do this to avoid having to allocate memory when no entries are
avaiable, this makes the code much simpler.

The functions where also renamed because to better reflect their
purpose.