Added MADT Parsing that also pages IOAPICs
This makes it possible to find and use IO-APICs.
Also disables the 8259 PIC.

The g_gsi_override_table uses indexes as an 8259 PIC irq line, such that
the Global System Interrupt is easy to find when for instance hooking the
PS/2 keyboard interrupt.

To have a single struct represent an ioapic it's convenient to also store
its virtual address and irq count. to get a virtual addres you need to
page the IOAPIC, and to get the irq cont you need to then read from the
paged location. So it's very convenient to do it all when reading the MADT
rather than having additional functions that fill in those values
afterwards.