Added Local APIC support
Uses xAPIC rather than x2APIC MSR registers, because it's backward
compatible and easy to access using structs instead of having to write
each register using a wrsmr wrapper.

Note that this is quite bare bones just to get PIC interrupts working,
there are no structs etc for ICR or any of the IVTs or timers. IMO that
should be written whenever it has to be used.