Setup local APIC timer interrupt
Setup the local APIC timer which will be used for context switching

Context switch handler:
If there is more than one process it pushes all registers, the rflags
and the process kernel stack ptr onto the process stack.
The process address space is unmapped and the process is put at the
back of the queue. The address space of the new process is mapped and
the process stack is switch so that all registers, the rflags and the
kernel stack can be restored. Before all return points an EOI is sent
to the local APIC.