Process starting
Added block which adds the first process to the process queue and jumps
to it. This block is disabled because we have no way of loading an ELF
binary right now.

The function start_user_process can be called after at least one process
has been started to add another process to the queue.